ファイル構成変更
