+++ msg2/WEB-INF/scr/Servlet/Sample2_3.java
@@ -1,3 +1,4 @@
+package Servlet;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.util.Random;
