+++ msg2/WEB-INF/scr/Servlet/Sample4_1.java
@@ -1,3 +1,4 @@
+package Servlet;
 import java.io.IOException;
 import java.io.PrintWriter;
