+++ msg2/WEB-INF/web.xml
@@ -8,7 +8,7 @@
 
   <servlet>
     <servlet-name>Sample</servlet-name>
-    <servlet-class>Sample</servlet-class>
+    <servlet-class>Servlet.Sample</servlet-class>
   </servlet>
   <servlet-mapping>
     <servlet-name>Sample</servlet-name>
@@ -17,7 +17,7 @@
 
   <servlet>
     <servlet-name>Sample2_1</servlet-name>
-    <servlet-class>Sample2_1</servlet-class>
+    <servlet-class>Servlet.Sample2_1</servlet-class>
   </servlet>
   <servlet-mapping>
     <servlet-name>Sample2_1</servlet-name>
@@ -26,7 +26,7 @@
 
   <servlet>
     <servlet-name>Sample2_2</servlet-name>
-    <servlet-class>Sample2_2</servlet-class>
+    <servlet-class>Servlet.Sample2_2</servlet-class>
   </servlet>
   <servlet-mapping>
     <servlet-name>Sample2_2</servlet-name>
@@ -35,7 +35,7 @@
 
   <servlet>
     <servlet-name>Sample2_3</servlet-name>
-    <servlet-class>Sample2_3</servlet-class>
+    <servlet-class>Servlet.Sample2_3</servlet-class>
   </servlet>
   <servlet-mapping>
     <servlet-name>Sample2_3</servlet-name>
@@ -44,7 +44,7 @@
 
   <servlet>
     <servlet-name>Sample3_1</servlet-name>
-    <servlet-class>Sample3_1</servlet-class>
+    <servlet-class>Servlet.Sample3_1</servlet-class>
   </servlet>
   <servlet-mapping>
     <servlet-name>Sample3_1</servlet-name>
@@ -53,7 +53,7 @@
 
   <servlet>
     <servlet-name>Sample4_1</servlet-name>
-    <servlet-class>Sample4_1</servlet-class>
+    <servlet-class>Servlet.Sample4_1</servlet-class>
   </servlet>
   <servlet-mapping>
     <servlet-name>Sample4_1</servlet-name>
